+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+Improved example
+Changed padding and margin from double to EdgeInsets
+Added BuildContext extension so now it's possible to take current breakpoint by doing context.breakpoint
+Changed BreakpointConfiguration to be const
+Changed BreakpointConfigurator, now it checks for breakpoints if they are in order
+
 ## 1.0.1
 
 Added example